Save, get and reset states

+++ getState.php
@@ -6,31 +6,33 @@
  * Time: 7:45 PM
  */
 
+ini_set("display_errors", 1);
 require_once "functions.php";
 
-$round = $_REQUEST["round"];
+$round = safeReturn($_REQUEST, "round");
 $fileName = getAuctionStateFile($round, true);
 
+echo $fileName;
 $auctionStateJson = "";
-if($fileName!=null)
-{
-    $auctionStateJson = file_get_contents($fileName);
-    if(isset($_REQUEST["reset"]))
-    {
-        for($i=$round+1;;$i++)
-        {
-            $fileName = getAuctionStateFile($i, true);
-            if($fileName==null)
-            {
-                break;
-            }
-            unlink($fileName);
-        }
-    }
-}
-else if($round==1)
+if($round==1 || !isStringSet($round))
 {
     $auctionStateJson = getInitialState();
     file_put_contents("auction_states/round1.txt", $auctionStateJson);
+}
+else if($fileName!=null)
+{
+    $auctionStateJson = file_get_contents($fileName);
+}
+if(isStringSet($round) && isset($_REQUEST["reset"]))
+{
+    for($i=$round+1;;$i++)
+    {
+        $fileName = getAuctionStateFile($i, true);
+        if($fileName==null)
+        {
+            break;
+        }
+        unlink($fileName);
+    }
 }
 echo $auctionStateJson;
